Player death

+++ Assets/Resources/Scripts/GameData/Units/Data/UnitData.cs
@@ -5,7 +5,7 @@ namespace GameJam
     [CreateAssetMenu(fileName = "Unit", menuName = "Data/UnitData", order = 2)]
     public class UnitData : ScriptableObject
     {
-        [field: SerializeField, Range(0, 40)] public float Health { get; private set; }
+        [field: SerializeField, Range(0, 40)] public int Health { get; private set; }
         [field: SerializeField, Range(0, 20)] public float Speed { get; private set; }
     }
 }
